TST: move to test_subclass.py
